Address PR Feedback

+++ MBBSEmu/DOS/Interrupts/IInterruptHandler.cs
@@ -2,7 +2,7 @@
 {
     public interface IInterruptHandler
     {
-        ushort Vector { get; }
+        byte Vector { get; }
         void Handle();
     }
 }
